fixed the config file casting to int

+++ OreVeins/src/defaultPackadge/BIF.java
@@ -67,7 +67,7 @@ public class BIF
 		}
 		else
 		{
-			probToSpawn = config.getInt(string+".SpawnProbability");
+			probToSpawn = config.getDouble(string+".SpawnProbability");
 		}
 	}
 }
